refactor: request 파일을 도메인별 폴더로 정리
