+++ src/main/java/com/sprint/mission/discodeit/dto/request/binarycontent/BinaryContentCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.request;
+package com.sprint.mission.discodeit.dto.request.binarycontent;
 
 public record BinaryContentCreateRequest(
         String fileName,
