+++ src/main/java/com/sprint/mission/discodeit/dto/request/channel/PrivateChannelCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.request;
+package com.sprint.mission.discodeit.dto.request.channel;
 
 import java.util.List;
 import java.util.UUID;
