+++ src/main/java/com/sprint/mission/discodeit/dto/request/channel/PublicChannelCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.request;
+package com.sprint.mission.discodeit.dto.request.channel;
 
 public record PublicChannelCreateRequest(
         String name,
