+++ src/main/java/com/sprint/mission/discodeit/dto/request/channel/PublicChannelUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.request;
+package com.sprint.mission.discodeit.dto.request.channel;
 
 public record PublicChannelUpdateRequest(
         String newName,
