+++ src/main/java/com/sprint/mission/discodeit/dto/request/message/MessageUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.request;
+package com.sprint.mission.discodeit.dto.request.message;
 
 public record MessageUpdateRequest(
         String content
