+++ src/main/java/com/sprint/mission/discodeit/dto/request/readstatus/ReadStatusUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.request;
+package com.sprint.mission.discodeit.dto.request.readstatus;
 
 import java.time.Instant;
