+++ src/main/java/com/sprint/mission/discodeit/dto/request/user/UserCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.request;
+package com.sprint.mission.discodeit.dto.request.user;
 
 public record UserCreateRequest(
         String username,
