+++ src/main/java/com/sprint/mission/discodeit/dto/request/user/UserLoginRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.request;
+package com.sprint.mission.discodeit.dto.request.user;
 
 import jakarta.validation.constraints.NotBlank;
