+++ src/main/java/com/sprint/mission/discodeit/dto/request/user/UserUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.request;
+package com.sprint.mission.discodeit.dto.request.user;
 
 public record UserUpdateRequest(
         String newUsername,
