+++ src/main/java/com/sprint/mission/discodeit/dto/request/userstatus/UserStatusCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.request;
+package com.sprint.mission.discodeit.dto.request.userstatus;
 
 import java.time.Instant;
 import java.util.UUID;
