+++ src/main/java/com/sprint/mission/discodeit/service/AuthService.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.dto.request.UserLoginRequest;
+import com.sprint.mission.discodeit.dto.request.user.UserLoginRequest;
 import com.sprint.mission.discodeit.entity.User;
 
 public interface AuthService {
